Define event and predicate methods on an anonymous module.
Rather than defining the event and predicate methods straight onto your
classes with class_eval, these methods are now added to an anonymous
module, which is then inserted into your class inheritance chain.

As a result, you should now be able to create methods on your class
with the same name as an event or predicate method, and simply use
super() to fire the original method, rather than needing to use
alias_method.